# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Initiative Tracker API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /initiatives/{id}/milestones:
    get:
      tags: [Milestones]
      operationId: listMilestones
      summary: List milestones for an initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of milestones.
          content:
            application/json:
              schema:
                type: object
                required: [milestones]
                properties:
                  milestones:
                    type: array
                    items:
                      $ref: '#/components/schemas/MilestoneWithAttachments'
    post:
      tags: [Milestones]
      operationId: createMilestone
      summary: Create a milestone (project-manager or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMilestoneRequest'
      responses:
        '201':
          description: Milestone created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Milestone'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /milestones/{id}:
    put:
      tags: [Milestones]
      operationId: updateMilestone
      summary: Update a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMilestoneRequest'
      responses:
        '200':
          description: Milestone updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Milestone'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Milestones]
      operationId: deleteMilestone
      summary: Delete a milestone (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /milestones/{id}/submit:
    post:
      tags: [Milestones]
      operationId: submitMilestone
      summary: Submit a milestone for approval (initiative owner / project-manager)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Milestone submitted for approval.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Milestone'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /milestones/{id}/approve:
    post:
      tags: [Milestones]
      operationId: approveMilestone
      summary: Approve a milestone (approver or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalRequest'
      responses:
        '200':
          description: Milestone approved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Milestone'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /milestones/{id}/reject:
    post:
      tags: [Milestones]
      operationId: rejectMilestone
      summary: Reject a milestone (approver or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalRequest'
      responses:
        '200':
          description: Milestone rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Milestone'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /milestones/{id}/attachments:
    post:
      tags: [Milestones]
      operationId: addAttachment
      summary: Add a file attachment to a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAttachmentRequest'
      responses:
        '201':
          description: Attachment added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileAttachment'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List all users (admin only)
      responses:
        '200':
          description: List of users with roles.
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserWithRole'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /users/{id}/role:
    put:
      tags: [Users]
      operationId: updateUserRole
      summary: Update a user's role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Role updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithRole'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    Milestone:
      type: object
      required: [id, initiativeId, title, status, weight, createdAt, updatedAt]
      properties:
        id:
          type: integer
        initiativeId:
          type: integer
        title:
          type: string
        description:
          type: [string, 'null']
        status:
          type: string
          enum: [pending, in_progress, submitted, approved, rejected]
        weight:
          type: number
          minimum: 0
          maximum: 100
        dueDate:
          type: [string, 'null']
          format: date
        approvedById:
          type: [string, 'null']
        approvedByName:
          type: [string, 'null']
        approvedAt:
          type: [string, 'null']
          format: date-time
        rejectionReason:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MilestoneWithAttachments:
      allOf:
        - $ref: '#/components/schemas/Milestone'
        - type: object
          required: [attachments]
          properties:
            attachments:
              type: array
              items:
                $ref: '#/components/schemas/FileAttachment'
    CreateMilestoneRequest:
      type: object
      required: [title, weight]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        weight:
          type: number
          minimum: 0
          maximum: 100
        dueDate:
          type: string
          format: date
    UpdateMilestoneRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        weight:
          type: number
          minimum: 0
          maximum: 100
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, in_progress, submitted, approved, rejected]
    ApprovalRequest:
      type: object
      properties:
        comment:
          type: string
    FileAttachment:
      type: object
      required: [id, milestoneId, fileName, objectPath, uploadedById, createdAt]
      properties:
        id:
          type: integer
        milestoneId:
          type: integer
        fileName:
          type: string
        contentType:
          type: [string, 'null']
        objectPath:
          type: string
        uploadedById:
          type: string
        uploadedByName:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    AddAttachmentRequest:
      type: object
      required: [fileName, objectPath]
      properties:
        fileName:
          type: string
          minLength: 1
        contentType:
          type: string
        objectPath:
          type: string
          minLength: 1
    UserWithRole:
      type: object
      required: [id, role, createdAt]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        role:
          type: string
          enum: [admin, project-manager, approver]
        createdAt:
          type: string
          format: date-time
    UpdateRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, project-manager, approver]