# lib/api-zod/src/generated/types/addAttachmentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Initiative Tracker API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddAttachmentRequest {
  /** @minLength 1 */
  fileName: string;
  contentType?: string;
  /** @minLength 1 */
  objectPath: string;
}