# lib/api-zod/src/generated/types/approvalRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Initiative Tracker API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ApprovalRequest {
  comment?: string;
}