# lib/api-zod/src/generated/types/createMilestoneRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Initiative Tracker API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMilestoneRequest {
  /** @minLength 1 */
  title: string;
  description?: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  weight: number;
  dueDate?: Date;
}