# lib/api-zod/src/generated/types/userWithRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Initiative Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserWithRoleRole } from "./userWithRoleRole";

export interface UserWithRole {
  id: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  role: UserWithRoleRole;
  createdAt: Date;
}